package/lz4: don't install programs by default
Commit dd6a0291f4f09abb6d34c070dd496d7867fbe525 ("package/lz4: make
program installation optional") make the installation of the lz4
programs optional. To preserve backward compatibility, this commit
added a "default y" to the BR2_PACKAGE_LZ4_PROGS.

While Buildroot definitely tries to preserve backward compatibility in
general, in this case, the backward compatibility would mean that
everbody would by default install those 220 KB of program, which are
most likely not needed for the majority of use cases. Installaing them
by default kind of breaks the Buildroot principle of minimalism.

In this specific case, it is believed that the principle of minimalism
wins over the principle of backward compatibility.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>