package/xen: bump version to 4.14.0
Bump the Xen package to the latest release. We can also remove a patch
to change the CFLAGS as it is no longer required.

The hash of the license file is changed, as ZPL-2.0 is no longer
listed as one of the licenses of 3rd party code imported into Xen.

Signed-off-by: Alistair Francis <alistair@alistair23.me>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>